Move value.ts back in the value directory
