+++ packages/alfa-css/src/index.ts
@@ -2,4 +2,3 @@ export * from "./syntax";
 export * from "./value";
 export * from "./unit";
 export * from "./calculation";
-export * from "./value/index"
