+++ packages/alfa-css/src/value/collection/list.ts
@@ -7,7 +7,7 @@ import { Parser } from "@siteimprove/alfa-parser";
 import { Slice } from "@siteimprove/alfa-slice";
 
 import { Token } from "../../syntax";
-import { Value } from "../../value";
+import { Value } from "../value";
 
 const { delimited, option, map, separatedList } = Parser;
