+++ packages/alfa-css/src/value/color/format.ts
@@ -1,4 +1,4 @@
-import { Value } from "../../value";
+import { Value } from "../value";
 
 /**
  * @internal
