+++ packages/alfa-css/src/value/image/gradient-linear.ts
@@ -7,7 +7,7 @@ import { Slice } from "@siteimprove/alfa-slice";
 import * as json from "@siteimprove/alfa-json";
 
 import { Token } from "../../syntax";
-import { Value } from "../../value";
+import { Value } from "../value";
 
 import { Angle } from "../numeric";
