+++ packages/alfa-css/src/value/image/gradient-radial.ts
@@ -9,7 +9,7 @@ import { Slice } from "@siteimprove/alfa-slice";
 import * as json from "@siteimprove/alfa-json";
 
 import { Token } from "../../syntax";
-import { Value } from "../../value";
+import { Value } from "../value";
 
 import { Length, Percentage } from "../numeric";
 import { Position } from "../position";
