+++ packages/alfa-css/src/value/image/image.ts
@@ -3,7 +3,7 @@ import { Parser } from "@siteimprove/alfa-parser";
 import { Slice } from "@siteimprove/alfa-slice";
 
 import { Token } from "../../syntax";
-import { Value } from "../../value";
+import { Value } from "../value";
 
 import { URL } from "./url";
 import { Gradient } from "./gradient";
