+++ packages/alfa-css/src/value/keyword.ts
@@ -4,7 +4,7 @@ import { Predicate } from "@siteimprove/alfa-predicate";
 import { Slice } from "@siteimprove/alfa-slice";
 
 import { Token } from "../syntax";
-import { Value } from "../value";
+import { Value } from "./value";
 
 const { map } = Parser;
 const { equals } = Predicate;
