+++ packages/alfa-css/src/value/numeric/number.ts
@@ -5,7 +5,7 @@ import { Slice } from "@siteimprove/alfa-slice";
 import { Math } from "../../calculation";
 import { Number as BaseNumber } from "../../calculation/numeric";
 
-import { Value } from "../../value";
+import { Value } from "../value";
 
 import { Numeric } from "./numeric";
 import { Token } from "../../syntax";
