+++ packages/alfa-css/src/value/shadow.ts
@@ -4,7 +4,7 @@ import { Err, Result } from "@siteimprove/alfa-result";
 import { Slice } from "@siteimprove/alfa-slice";
 
 import { Token } from "../syntax";
-import { Value } from "../value";
+import { Value } from "./value";
 
 import { Color } from "./color";
 import { Keyword } from "./keyword";
