+++ packages/alfa-css/src/value/shape/basic-shape.ts
@@ -1,4 +1,4 @@
-import { Value } from "../../value";
+import { Value } from "../value";
 
 /**
  * @internal
