+++ packages/alfa-css/src/value/string.ts
@@ -3,7 +3,7 @@ import { Parser } from "@siteimprove/alfa-parser";
 import { Slice } from "@siteimprove/alfa-slice";
 
 import { Token } from "../syntax";
-import { Value } from "../value";
+import { Value } from "./value";
 
 const { map } = Parser;
