+++ packages/alfa-css/src/value/transform/function.ts
@@ -1,4 +1,4 @@
-import { Value } from "../../value";
+import { Value } from "../value";
 
 /**
  * @internal
